requirements.txt
bambulabs_api==2.6.3
blinker==1.9.0
click==8.2.1
Flask==3.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
paho-mqtt==2.1.0
pillow==11.3.0
python-dotenv==1.0.0
Werkzeug==3.0.1
